First service created

+++ src/main/java/ing/yisus/apihotelera/repository/HotelRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HotelRepository extends JpaRepository<HotelEntity, Long> {}
+public interface HotelRepository extends JpaRepository<HotelEntity, Integer> {}
